final 3.8, 3.9

+++ src/main/java/org/iesvdm/dto/ComercialDTO2.java
@@ -2,6 +2,7 @@ package org.iesvdm.dto;
 
 import lombok.Data;
 import org.iesvdm.modelo.Comercial;
+import java.math.BigDecimal;
 
 @Data
 public class ComercialDTO2 extends Comercial{
@@ -10,7 +11,7 @@ public class ComercialDTO2 extends Comercial{
     private int pedidosUltimoSemestre;
     private int pedidosUltimoAnio;
     private int pedidosUltimoLustro;
-    public ComercialDTO2(int id, String nombre, String apellido1, String apellido2, float comision,
+    public ComercialDTO2(int id, String nombre, String apellido1, String apellido2, BigDecimal comision,
                          int numeroPedidosPorCliente, int pedidosUltimoTrimestre,
                          int pedidosUltimoSemestre, int pedidosUltimoAnio, int pedidosUltimoLustro) {
         super(id, nombre, apellido1, apellido2, comision);
